Update checker.sh
gsutil ls is returning nothing for an existing, valid bucket without any files. Assumption is this is a new file created for this purpose, so it shouldn't have anything. And if the bucket does not actually exist, the gsutil command does return an error like:
BucketNotFoundException: 404 gs://rabia-tengo-fake-bucket bucket does not exist.

@fawix please review and merge